Updated Launcher script, added new scripts for lobby UI and added simple lobby UI. Created new folders in project
Launcher script now have 3 new events OnConnectedToRoomEvent this event is invoked when player is connected to the room, OnConnectingStartedEvent this one is invoked on start of connecting proces. OnDisconnectedEvent this event is invoked when player is disconnecting from the room, but now there are no disconnecting feature implemented.

Created new folders like Prefabs, UI, and inside of Code Base folder there are two new folders Player and UI.

Added new scripts PlayerNameInputField and UIController

PlayerNameInputField is setting player name on start from the player prefs, if there are no saved name, then input field text is setted to 'Enter your nickname'. Also this script has method that called SetPlayerName this method is setting nickname for player from the input field if this input field is not empty or null.

UI Controller controlls the UI gameObjects, setting them to active and inactive depending on event that invoked in the launcher